backend/requirements.txt
Django==5.2.13
django-mongodb-backend==5.2.3
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
pymongo==4.16.0
